feat(professor): Add Adedonha game manager component
Introduces a new component `ProfessorAdedonhaManager` for professors to create and manage "Adedonha" (also known as "Stop!") games. This feature allows professors to select a course, define game categories, set a time limit, and start games. It also includes logic to automatically reset the game if the selected course is deleted. A new `CubeIcon` has also been added to support this new component.